Exclude PRs whose state is not "open"
The search we do should achieve this for us, but for some reason we were
getting back a PR that was closed.  I suspect it's some search index thing
on the github side that is broken for this PR.  It makes sense to filter
them out if we can though.