Filter out draft PRs
The draft attribute is in preview so we need to change our accept-header
to make sure we get it in the payload.  The draft attribute is also only
available on the full pull request API response, not the minimal version
we get from doing a search.  We already make this request to calculate the
comment count, so I've moved it out of there and into the
`pull_requests_from_github` method, where we map over the search results
and make the full pr request to github and return those instead of the
simpler search results.